chore: processing temperature
~ now by default temperature is stored as celcius value removing the
  need for extra computation later

+++ main/include/Motion.h
@@ -27,7 +27,7 @@ typedef struct {
         int16_t X, Y, Z;
     } Gyro;
 
-    int16_t Temperature;
+    double Temperature;
 } MotionValues;
